mejoras en toolbar, agrego un rectangulo gris en main activity, emprolijo algunos archivos, extraigo funcionalidad utilitaria a su propia clase
