+++ app/src/main/AndroidManifest.xml
@@ -10,15 +10,14 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Base.Theme.ParcialTP3"
+        android:theme="@style/Theme.ParcialTP3"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="false" />
         <activity
             android:name=".LoginActivity"
-            android:exported="true"
-            android:theme="@style/Theme.ParcialTP3.NoActionBar" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
