+++ app/src/main/java/com/example/parcialtp3/LoginActivity.kt
@@ -1,17 +1,12 @@
 package com.example.parcialtp3
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Button
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
-
-
-
 
     }
 }
