+++ app/src/main/res/drawable/atras.xml
@@ -8,6 +8,6 @@
       android:strokeLineJoin="round"
       android:strokeWidth="2.5"
       android:fillColor="#00000000"
-      android:strokeColor="#343B42"
+      android:strokeColor="#ffffff"
       android:strokeLineCap="round"/>
 </vector>
