+++ app/src/main/res/drawable/button_rounded_background.xml
@@ -1,4 +1,4 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/white" />
+    <solid android:color="@color/grey" />
     <corners android:radius="28dp" />
 </shape>
