+++ app/src/main/res/layout/activity_main.xml
@@ -8,10 +8,25 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout2"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
+
+        <View
+            android:id="@+id/rectangle_3"
+            android:layout_width="390dp"
+            android:layout_height="734dp"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentTop="true"
+            android:background="@drawable/rectangle_3"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.94" />
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/appbar"
@@ -24,7 +39,7 @@
             app:layout_constraintTop_toTopOf="parent">
 
             <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbarx"
+                android:id="@+id/toolbar"
                 android:layout_width="match_parent"
                 android:layout_height="?attr/actionBarSize">
 
@@ -33,13 +48,6 @@
                     android:layout_height="match_parent"
                     android:gravity="center_vertical"
                     android:orientation="horizontal">
-
-                    <ImageView
-                        android:id="@+id/icon"
-                        android:layout_width="40dp"
-                        android:layout_height="40dp"
-                        android:scaleType="centerInside"
-                        android:src="@drawable/hamburger" />
 
                     <TextView
                         android:id="@+id/titleToolbar"
@@ -59,7 +67,8 @@
                         android:layout_marginTop="3dp"
                         android:layout_marginEnd="20dp"
                         android:scaleType="centerCrop"
-                        android:src="@drawable/avatar_car" />
+                        android:src="@drawable/avatar_car"
+                        android:contentDescription="Av" />
 
                 </LinearLayout>
             </androidx.appcompat.widget.Toolbar>
@@ -82,15 +91,18 @@
         </com.google.android.material.bottomnavigation.BottomNavigationView>
 
         <androidx.drawerlayout.widget.DrawerLayout
+
             android:id="@+id/drawer_layout"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_marginTop="60dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
 
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.0"
             tools:context=".MainActivity">
 
             <com.google.android.material.navigation.NavigationView
@@ -101,19 +113,17 @@
                 android:fitsSystemWindows="true"
                 app:menu="@menu/drawer_menu" />
 
+            <androidx.fragment.app.FragmentContainerView
+                android:id="@+id/nav_host"
+                android:name="androidx.navigation.fragment.NavHostFragment"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                app:defaultNavHost="true"
+                app:navGraph="@navigation/nav_main" />
+
         </androidx.drawerlayout.widget.DrawerLayout>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/nav_host"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:defaultNavHost="true"
-        app:layout_anchor="@+id/constraintLayout2"
-        app:layout_anchorGravity="center"
-        app:navGraph="@navigation/nav_main" />
 
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
