+++ app/src/main/res/layout/fragment_login.xml
@@ -92,13 +92,16 @@
         android:layout_marginStart="147dp"
         android:layout_marginTop="50dp"
         android:layout_marginEnd="148dp"
-        android:fontFamily="@font/poppins_bold"
+        android:layout_marginBottom="303dp"
         android:background="@drawable/button_rounded_background"
+        android:fontFamily="@font/poppins_bold"
         android:text="@string/btn_ingresar"
         android:textColor="#FFFFFF"
         android:textSize="19sp"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtOlvido" />
+        app:layout_constraintTop_toBottomOf="@+id/txtOlvido"
+        app:strokeColor="#212121" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
