+++ app/src/main/res/navigation/nav_login.xml
@@ -12,7 +12,8 @@
         tools:layout="@layout/fragment_login" >
         <action
             android:id="@+id/action_loginFragment_to_mainActivity"
-            app:destination="@id/mainActivity" />
+            app:destination="@id/mainActivity"
+            app:launchSingleTop="false" />
     </fragment>
     <activity
         android:id="@+id/mainActivity"
