+++ app/src/main/res/values/themes.xml
@@ -1,26 +1,11 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Estilo base para la aplicación con barra de acción -->
-    <style name="Base.Theme.ParcialTP3" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
-        <!-- Personaliza aquí tu tema con barra de acción -->
+    <!-- Tema oscuro -->
+    <style name="Theme.ParcialTP3.Dark" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <item name="colorPrimary">@color/black</item>
-        <!-- Otros atributos y personalizaciones -->
     </style>
 
-    <!-- Estilo para la aplicación con barra de acción -->
-    <style name="Theme.ParcialTP3" parent="Base.Theme.ParcialTP3">
-        <!-- Personaliza aquí tu tema con barra de acción -->
-    </style>
-
-    <!-- Estilo base para la aplicación sin barra de acción -->
-    <style name="Base.Theme.ParcialTP3.NoActionBar" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <!-- Personaliza aquí tu tema sin barra de acción -->
+    <!-- Tema standard -->
+    <style name="Theme.ParcialTP3" parent="Theme.MaterialComponents.Light.NoActionBar">
         <item name="colorPrimary">@color/grey</item>
-        <!-- Otros atributos y personalizaciones -->
     </style>
-
-    <!-- Estilo para la aplicación sin barra de acción -->
-    <style name="Theme.ParcialTP3.NoActionBar" parent="Base.Theme.ParcialTP3.NoActionBar">
-        <!-- Personaliza aquí tu tema sin barra de acción -->
-    </style>
-
 </resources>
